Add files via upload
In this, I have made a python program for rollercoster entry. Only the people who are more than 120 cms tall can ride the rollercoster. Children under the age of 12 have to pay $5, teens from the age of 13 to 18 have to pay $7 and people above the age of the 18 have to pay $12.